fix weird things: working in dockerized localhost

+++ backend/src/main/kotlin/com/lucasalfare/flrefs/main/routes/ClearAllRoute.kt
@@ -30,11 +30,12 @@ fun Route.clearAllItemsRoute() {
           else ""
 
         SchemaUtils.listTables().forEach {
-          exec("DROP TABLE IF EXISTS $it$cascade")
+          if (it != Users.tableName) {
+            exec("DROP TABLE IF EXISTS $it$cascade")
+          }
         }
 
         SchemaUtils.createMissingTablesAndColumns(
-          Users,
           ImagesInfos,
           ImagesUrls
         )
